Rename Cluster service to Control
This will pave the way for adding a top-level Cluster object that's
manipulated by the Control service.
