+++ manager/controlapi/manager_test.go
@@ -1,4 +1,4 @@
-package clusterapi
+package controlapi
 
 import (
 	"fmt"
