+++ manager/controlapi/network.go
@@ -1,4 +1,4 @@
-package clusterapi
+package controlapi
 
 import (
 	"net"
