+++ manager/controlapi/node_test.go
@@ -1,4 +1,4 @@
-package clusterapi
+package controlapi
 
 import (
 	"testing"
