+++ manager/controlapi/server.go
@@ -1,4 +1,4 @@
-package clusterapi
+package controlapi
 
 import (
 	"errors"
